Extract local_config callback
